requirements.txt
chromadb==0.3.27
langchain==0.0.230
openai==0.27.8
tiktoken==0.4.0
streamlit==1.24.1
pinecone-client==2.2.2
wheel==0.41.0
unstructured==0.8.1